Handle issues with lightcurve analysis
Omit bins falling within non-science periods, which have no data.

Handle specific RuntimeErrors. There had been a specific RuntimeError
affecting the analysis as a whole, not a specific bin, with the result that
the exception was repeatedly caught and all analysis output deleted.

Remove the last-modified directory, not the last-numbered directory,
since with repeated iterations the problematic directory might not be the
last one.

Set the process limit higher to avoid crashes due to too many open processes.